Speed up celery tests by using worker fixture
